Create SingleFRAPSummarizer.ipynb
This Jupyter notebook loads up the different FRAP.csvs from the macro and creates an analysis .csv in the same folder. It also does Phair bleach correction and normalizes the FRAP curve 0-1. It then plots the curves and saves them in the same folder as the original csvs. Finally it summarizes all the normalized FRAP curves sorts them by reference method and creates a final summary csv. It is well adviced to check the curves for irregularities. The macro may have made some ill judgement concerning the bleach